Allow warning about duplicates instead of erroring
In a workflow with multiple inputs like one with local private sequences
and sequences from GISAID, it is possible for the same strain to have
different local and GISAID sequences. When this happens, we want to
allow users to prefer their local sequences over the GISAID versions
during the combine-and-dedup step without throwing an error and
requiring manual modification of GISAID data.

This commit adds a `--warn-about-duplicates` flag to the
combine-and-dedup script that allows the script to exit with a normal
error code while still printing a warning about which strains had
duplicate sequences. This commit also adds a configuration parameter for
the combine-and-dedup rule so users can turn on warnings, as they wish.